requirements.txt
appdirs==1.4.3
autopep8==2.0.2
beautifulsoup4==4.11.2
CacheControl==0.12.6
cachetools==5.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
db-dtypes==1.0.5
distlib==0.3.0
distro==1.4.0
google-api-core==2.11.0
google-auth==2.16.2
google-cloud-bigquery==3.7.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
numpy==1.24.2
packaging==20.3
pandas==1.5.3
pep517==0.8.2
polars==0.16.14
praw==7.7.0
prawcore==2.3.0
progress==1.5
proto-plus==1.22.2
protobuf==4.22.1
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pyparsing==2.4.6
python-dateutil==2.8.2
pytoml==0.1.21
pytz==2022.7.1
requests==2.22.0
retrying==1.3.3
rsa==4.9
six==1.14.0
soupsieve==2.4
tomli==2.0.1
typing-extensions==4.5.0
update-checker==0.18.0
urllib3==1.25.8
webencodings==0.5.1
websocket-client==1.5.1
